app/requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
dialogflow==0.7.2
Flask==1.1.1
Flask-Login==0.4.1
Flask-MySQL==1.4.0
Flask-SQLAlchemy==2.4.1
google-api-core==1.14.3
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.24.3
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.3
Pillow==6.2.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.10
http://download.pytorch.org/whl/cpu/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
torchvision==0.4.1
urllib3==1.25.6
Werkzeug==0.16.0
